Fixed multiple instances of being able to cheat.
Added field to be able to type in guesses that allows for single letters or a whole word.

Added a richer set of words that are loaded from a wordlist txt file.

Fixed variables that were changing types based on boolean conditionals.

Added comments to make the code more clear.